refactor links parsing
