+++ test/membrane/core/parent/child_life_controller/link_handler_test.exs
@@ -6,7 +6,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
   alias Membrane.ChildEntry
 
   alias Membrane.Core.Message
-  alias Membrane.Core.Parent.Link
+  alias Membrane.Core.Parent.LinkParser
   alias Membrane.Core.Parent.ChildLifeController.LinkHandler
   alias Membrane.LinkError
   alias Membrane.Pad
@@ -44,13 +44,13 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
 
   describe "resolve links" do
     test "should work for static pads" do
-      {:ok, links} = Link.from_spec([link(:a) |> to(:b)])
+      {:ok, links} = LinkParser.parse_links([link(:a) |> to(:b)])
       resolved_links = LinkHandler.resolve_links(links, get_state(TestFilter))
       endpoints(resolved_links) |> Enum.each(&assert &1.pad_ref == &1.pad_spec)
     end
 
     test "should work for dynamic pads" do
-      {:ok, links} = Link.from_spec([link(:a) |> to(:b)])
+      {:ok, links} = LinkParser.parse_links([link(:a) |> to(:b)])
 
       resolved_links = LinkHandler.resolve_links(links, get_state(TestDynamicPadFilter))
 
@@ -60,7 +60,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
       end)
 
       {:ok, links} =
-        Link.from_spec([
+        LinkParser.parse_links([
           link(:a) |> via_out(Pad.ref(:output, :x)) |> via_in(Pad.ref(:input, :y)) |> to(:b)
         ])
 
@@ -71,7 +71,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
     end
 
     test "should work for bin static pads" do
-      {:ok, links} = Link.from_spec([link_bin_input() |> to(:a) |> to_bin_output()])
+      {:ok, links} = LinkParser.parse_links([link_bin_input() |> to(:a) |> to_bin_output()])
 
       resolved_links = LinkHandler.resolve_links(links, get_state(TestFilter))
       endpoints(resolved_links) |> Enum.each(&assert &1.pad_ref == &1.pad_spec)
@@ -83,7 +83,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
 
     test "should work for bin dynamic pads" do
       {:ok, links} =
-        Link.from_spec([
+        LinkParser.parse_links([
           link_bin_input(Pad.ref(:input, :x)) |> to(:a) |> to_bin_output(Pad.ref(:output, :y))
         ])
 
@@ -103,7 +103,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
     end
 
     test "should fail when trying to link non-existent child" do
-      {:ok, links} = Link.from_spec([link(:a) |> to(:m)])
+      {:ok, links} = LinkParser.parse_links([link(:a) |> to(:m)])
 
       assert_raise LinkError, ~r/child :m does not exist/i, fn ->
         LinkHandler.resolve_links(links, get_state(TestFilter))
@@ -111,7 +111,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
     end
 
     test "should fail when trying to link non-existent pad" do
-      {:ok, links} = Link.from_spec([link(:a) |> via_out(:x) |> to(:b)])
+      {:ok, links} = LinkParser.parse_links([link(:a) |> via_out(:x) |> to(:b)])
 
       assert_raise LinkError, ~r/child :a does not have pad :x/i, fn ->
         LinkHandler.resolve_links(links, get_state(TestFilter))
@@ -119,7 +119,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
     end
 
     test "should fail when trying to pass dynamic pad ref to a static pad" do
-      {:ok, links} = Link.from_spec([link(:a) |> via_out(Pad.ref(:output, :x)) |> to(:b)])
+      {:ok, links} = LinkParser.parse_links([link(:a) |> via_out(Pad.ref(:output, :x)) |> to(:b)])
 
       assert_raise LinkError,
                    ~r/dynamic pad ref .*membrane.pad.*:output.* passed for static pad of child :a/i,
@@ -129,7 +129,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
     end
 
     test "should fail when trying to link non-existent bin pad" do
-      {:ok, links} = Link.from_spec([link_bin_input(:x) |> to(:b)])
+      {:ok, links} = LinkParser.parse_links([link_bin_input(:x) |> to(:b)])
 
       assert_raise LinkError, ~r/bin :my_bin does not have pad :x/i, fn ->
         LinkHandler.resolve_links(links, get_state(TestFilter))
@@ -137,7 +137,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
     end
 
     test "should fail when trying to link dynamic bin pad and passed only name" do
-      {:ok, links} = Link.from_spec([link_bin_input(:input) |> to(:b)])
+      {:ok, links} = LinkParser.parse_links([link_bin_input(:input) |> to(:b)])
 
       assert_raise LinkError,
                    ~r/exact reference not passed when linking dynamic bin pad :input/i,
@@ -150,7 +150,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
     end
 
     test "should fail when trying to link dynamic bin pad that is not externally linked yet" do
-      {:ok, links} = Link.from_spec([link_bin_input(Pad.ref(:input, :x)) |> to(:b)])
+      {:ok, links} = LinkParser.parse_links([link_bin_input(Pad.ref(:input, :x)) |> to(:b)])
 
       assert_raise LinkError,
                    ~r/linking dynamic bin pad .*membrane.pad.*:input.* not .* externally linked/i,
@@ -163,7 +163,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
     end
 
     test "should fail when trying to pass dynamic pad ref to a static bin pad" do
-      {:ok, links} = Link.from_spec([link(:a) |> to_bin_output(Pad.ref(:output, :x))])
+      {:ok, links} = LinkParser.parse_links([link(:a) |> to_bin_output(Pad.ref(:output, :x))])
 
       assert_raise LinkError,
                    ~r/dynamic pad ref .*membrane.pad.*:output.* passed for static pad of bin :my_bin/i,
@@ -198,7 +198,7 @@ defmodule Membrane.Core.Parent.ChildLifeController.LinkHandlerTest do
 
     {:ok, pid} = GenServer.start_link(Proxy, self())
     state = get_state(TestFilter, pid: pid)
-    {:ok, links} = Link.from_spec([link(:a) |> to(:b)])
+    {:ok, links} = LinkParser.parse_links([link(:a) |> to(:b)])
     resolved_links = LinkHandler.resolve_links(links, state)
     assert {:ok, _state} = LinkHandler.link_children(resolved_links, state)
     assert_receive Message.new(:handle_link, _args)
